Remove useless cast on AudioDestination::maxChannelCount().
Reviewed by: Marcelo Lira

+++ Source/WebCore/platform/audio/nix/AudioDestinationNix.cpp
@@ -116,7 +116,7 @@ float AudioDestination::hardwareSampleRate()
 
 unsigned long AudioDestination::maxChannelCount()
 {
-    return static_cast<float>(Nix::Platform::current()->audioHardwareOutputChannels());
+    return Nix::Platform::current()->audioHardwareOutputChannels();
 }
 
 void AudioDestinationNix::render(const Nix::Vector<float*>& sourceData, const Nix::Vector<float*>& audioData, size_t numberOfFrames)
